feat(dns): switch DoH→plain TCP DNS, drive upstreams from Settings
The trusted DNS path now speaks RFC 1035 length-prefixed messages over
mihomo_tunnel::tcp::handle_tcp instead of HTTPS — same in-process
dispatch, so the proxy chain (selector → SOCKS5 outbound → upstream) is
honored. The cache, single-flight dedup, and persistent doh-cache.redb
all carry over; the redb filename is preserved for compat with shipped
TestFlight builds.

The trusted-resolver upstream pool is no longer read from config.yaml.
A new FFI surface (`meow_dns_set_upstreams`) takes a comma-/whitespace-
separated CSV from the iOS Settings view and stashes it before
`meow_tun_start`. Empty / unset / parse-failed → built-in defaults
(1.1.1.1 / 8.8.8.8). The China-DNS split-horizon path still reads
`dns.china_nameserver` from YAML.

Drops hyper / hyper-util / http-body-util / tokio-rustls / bytes deps;
keeps rustls (engine::install_tls_provider still wires the ring
provider). Cargo.lock shrinks by 6 lines.
