+++ MeowShared/Sources/MeowModels/Preferences.swift
@@ -4,7 +4,7 @@ import Foundation
 public enum PreferenceKey {
     public static let mixedPort = "com.meow.mixedPort"
     public static let localDnsPort = "com.meow.localDnsPort"
-    public static let dohServer = "com.meow.dohServer"
+    public static let dnsServers = "com.meow.dnsServers"
     public static let logLevel = "com.meow.logLevel"
     public static let allowLan = "com.meow.allowLan"
     public static let ipv6 = "com.meow.ipv6"
@@ -18,7 +18,9 @@ public enum PreferenceKey {
 public enum PreferenceDefaults {
     public static let mixedPort: Int = 7890
     public static let localDnsPort: Int = 1053
-    public static let dohServer: String = ""
+    /// Comma-separated list of plain-TCP DNS upstreams. Empty → the Rust
+    /// engine falls back to its built-in defaults (1.1.1.1 / 8.8.8.8).
+    public static let dnsServers: String = ""
     public static let logLevel: String = "info"
     public static let allowLan: Bool = false
     public static let ipv6: Bool = false
@@ -28,7 +30,7 @@ public enum PreferenceDefaults {
 public struct Preferences: Sendable {
     public var mixedPort: Int
     public var localDnsPort: Int
-    public var dohServer: String
+    public var dnsServers: String
     public var logLevel: String
     public var allowLan: Bool
     public var ipv6: Bool
@@ -36,14 +38,14 @@ public struct Preferences: Sendable {
     public init(
         mixedPort: Int = PreferenceDefaults.mixedPort,
         localDnsPort: Int = PreferenceDefaults.localDnsPort,
-        dohServer: String = PreferenceDefaults.dohServer,
+        dnsServers: String = PreferenceDefaults.dnsServers,
         logLevel: String = PreferenceDefaults.logLevel,
         allowLan: Bool = PreferenceDefaults.allowLan,
         ipv6: Bool = PreferenceDefaults.ipv6,
     ) {
         self.mixedPort = mixedPort
         self.localDnsPort = localDnsPort
-        self.dohServer = dohServer
+        self.dnsServers = dnsServers
         self.logLevel = logLevel
         self.allowLan = allowLan
         self.ipv6 = ipv6
@@ -57,7 +59,7 @@ public struct Preferences: Sendable {
         if defaults.object(forKey: PreferenceKey.localDnsPort) != nil {
             prefs.localDnsPort = defaults.integer(forKey: PreferenceKey.localDnsPort)
         }
-        prefs.dohServer = defaults.string(forKey: PreferenceKey.dohServer) ?? PreferenceDefaults.dohServer
+        prefs.dnsServers = defaults.string(forKey: PreferenceKey.dnsServers) ?? PreferenceDefaults.dnsServers
         prefs.logLevel = defaults.string(forKey: PreferenceKey.logLevel) ?? PreferenceDefaults.logLevel
         if defaults.object(forKey: PreferenceKey.allowLan) != nil {
             prefs.allowLan = defaults.bool(forKey: PreferenceKey.allowLan)
@@ -71,7 +73,7 @@ public struct Preferences: Sendable {
     public func save(to defaults: UserDefaults) {
         defaults.set(mixedPort, forKey: PreferenceKey.mixedPort)
         defaults.set(localDnsPort, forKey: PreferenceKey.localDnsPort)
-        defaults.set(dohServer, forKey: PreferenceKey.dohServer)
+        defaults.set(dnsServers, forKey: PreferenceKey.dnsServers)
         defaults.set(logLevel, forKey: PreferenceKey.logLevel)
         defaults.set(allowLan, forKey: PreferenceKey.allowLan)
         defaults.set(ipv6, forKey: PreferenceKey.ipv6)
